requirements.txt
colorama==0.3.9
coverage==4.5.1
Flask==1.0.2
Flask-JWT==0.3.2
Flask-RESTful==0.3.6
PyJWT==1.4.2
pylint==1.9.2
pytest==3.6.1
coverage-badge==0.2.0
gunicorn==19.8.1
psycopg2==2.7.5